chore: rewrote all api methods to pages.py
