+++ fossunited/chapters/web_template/foss_clubs_home_page/foss_clubs_home_page.html
@@ -244,7 +244,7 @@
             function getClubs(query) {
                 return new Promise((resolve, reject) => {
                     frappe.call({
-                        method: "fossunited.chapters.web_template.foss_clubs_home_page.foss_clubs_home_page.search_foss_club",
+                        method: "fossunited.api.pages.search_foss_club",
                         args: { query },
                         callback: (r) => {
                             resolve(r.message);
